requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
cachetools==5.3.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.104.1
google-auth==2.26.1
h11==0.14.0
hashids==1.3.1
idna==3.6
jose==1.0.0
Mako==1.3.0
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.3
pandas==2.1.4
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.4
uvicorn==0.24.0.post1